# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/trip": {
            "get": {
                "tags": [
                    "Voyage"
                ],
                "summary": "Obtenir la liste des voyages",
                "operationId": "cadf5c68469d51c35eaca2a49b49acb0",
                "responses": {
                    "200": {
                        "description": "Liste des voyages récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Trip"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur interne"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Voyage"
                ],
                "summary": "Créer un nouveau voyage",
                "operationId": "881dc7b53d5d143c7d8440a15d4c06ab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "starting_point",
                                    "ending_point",
                                    "starting_at",
                                    "available_places",
                                    "price",
                                    "user_id"
                                ],
                                "properties": {
                                    "starting_point": {
                                        "type": "string",
                                        "example": "Paris"
                                    },
                                    "ending_point": {
                                        "type": "string",
                                        "example": "Londres"
                                    },
                                    "starting_at": {
                                        "type": "string",
                                        "example": "2024-08-01T10:00:00Z"
                                    },
                                    "available_places": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Voyage créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Trip"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requête incorrecte"
                    },
                    "500": {
                        "description": "Erreur serveur interne"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/trip/{id}": {
            "get": {
                "tags": [
                    "Voyage"
                ],
                "summary": "Obtenir un voyage par ID",
                "operationId": "2110cc122ffe7d648ff3e141763b85b6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du voyage",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Voyage récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Trip"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Voyage non trouvé"
                    },
                    "500": {
                        "description": "Erreur serveur interne"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Voyage"
                ],
                "summary": "Mettre à jour un voyage",
                "operationId": "d3de21a63aca66d337bc467a52748e98",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du voyage",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "starting_point": {
                                        "type": "string",
                                        "example": "Paris"
                                    },
                                    "ending_point": {
                                        "type": "string",
                                        "example": "Londres"
                                    },
                                    "starting_at": {
                                        "type": "string",
                                        "example": "2024-08-01T10:00:00Z"
                                    },
                                    "available_places": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Voyage mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Trip"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requête incorrecte"
                    },
                    "404": {
                        "description": "Voyage non trouvé"
                    },
                    "500": {
                        "description": "Erreur serveur interne"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Voyage"
                ],
                "summary": "Supprimer un voyage",
                "operationId": "df8984a530e8a1ef6175f14251edd2f3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du voyage",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Voyage supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Voyage supprimé avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Voyage non trouvé"
                    },
                    "500": {
                        "description": "Erreur serveur interne"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/trips": {
            "get": {
                "tags": [
                    "Trips"
                ],
                "summary": "Rechercher des trajets",
                "operationId": "9ffcbfef8d645c0f7505d3a50bf2b8af",
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "description": "Point de départ du trajet",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "end",
                        "in": "query",
                        "description": "Point d'arrivée du trajet",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Date du trajet",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des trajets récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Trip"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun trajet trouvé"
                    },
                    "500": {
                        "description": "Erreur serveur interne"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Trip": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "starting_point": {
                        "type": "string",
                        "example": "Paris"
                    },
                    "ending_point": {
                        "type": "string",
                        "example": "Lyon"
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-08-01"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-29T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-29T12:34:56Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Bearer token authentication",
                "scheme": "bearer"
            },
            "bearerAuth": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}